requirements.txt
torch
torchvision
roma
gradio
einops
trimesh
scipy
matplotlib
pillow
tqdm
numpy
opencv-python
